+++ protobuf/lib/src/protobuf/field_set.dart
@@ -70,8 +70,8 @@ class _FieldSet {
   int? get _memoizedHashCode =>
       _frozenState is int ? _frozenState as int : null;
 
-  // Maps a oneof decl index to the tag number which is currently set. If the
-  // index is not present, the oneof field is unset.
+  /// Maps a `oneof` field index to the tag number which is currently set. If
+  /// the index is not present, the oneof field is unset.
   final Map<int, int>? _oneofCases;
 
   _FieldSet(this._message, BuilderInfo meta, this._eventPlugin)
@@ -98,16 +98,8 @@ class _FieldSet {
   /// The FieldInfo for each non-extension field in tag order.
   Iterable<FieldInfo> get _infosSortedByTag => _meta.sortedByTag;
 
-  /// Returns true if we should send events to the plugin.
-  bool get _hasObservers => _eventPlugin != null && _eventPlugin!.hasObservers;
-
-  bool get _hasExtensions => _extensions != null;
-  bool get _hasUnknownFields => _unknownFields != null;
-
-  _ExtensionFieldSet _ensureExtensions() {
-    if (!_hasExtensions) _extensions = _ExtensionFieldSet(this);
-    return _extensions!;
-  }
+  _ExtensionFieldSet _ensureExtensions() =>
+      _extensions ??= _ExtensionFieldSet(this);
 
   UnknownFieldSet _ensureUnknownFields() {
     if (_unknownFields == null) {
@@ -138,8 +130,7 @@ class _FieldSet {
   FieldInfo? _getFieldInfoOrNull(int tagNumber) {
     var fi = _nonExtensionInfo(_meta, tagNumber);
     if (fi != null) return fi;
-    if (!_hasExtensions) return null;
-    return _extensions!._getInfoOrNull(tagNumber);
+    return _extensions?._getInfoOrNull(tagNumber);
   }
 
   void _markReadOnly() {
@@ -161,13 +152,9 @@ class _FieldSet {
         }
       }
     }
-    if (_hasExtensions) {
-      _ensureExtensions()._markReadOnly();
-    }
 
-    if (_hasUnknownFields) {
-      _ensureUnknownFields()._markReadOnly();
-    }
+    _extensions?._markReadOnly();
+    _unknownFields?._markReadOnly();
   }
 
   void _ensureWritable() {
@@ -190,10 +177,11 @@ class _FieldSet {
       if (value != null) return value;
       return _getDefault(fi);
     }
-    if (_hasExtensions) {
-      var fi = _extensions!._getInfoOrNull(tagNumber);
+    final extensions = _extensions;
+    if (extensions != null) {
+      var fi = extensions._getInfoOrNull(tagNumber);
       if (fi != null) {
-        return _extensions!._getFieldOrDefault(fi);
+        return extensions._getFieldOrDefault(fi);
       }
     }
     throw ArgumentError('tag $tagNumber not defined in $_messageName');
@@ -220,39 +208,43 @@ class _FieldSet {
   /// Works for both repeated and non-repeated fields.
   dynamic _getFieldOrNull(FieldInfo fi) {
     if (fi.index != null) return _values[fi.index!];
-    if (!_hasExtensions) return null;
-    return _extensions!._getFieldOrNull(fi as Extension<dynamic>);
+    return _extensions?._getFieldOrNull(fi as Extension<dynamic>);
   }
 
   bool _hasField(int tagNumber) {
     var fi = _nonExtensionInfo(_meta, tagNumber);
     if (fi != null) return _$has(fi.index!);
-    if (!_hasExtensions) return false;
-    return _extensions!._hasField(tagNumber);
+    return _extensions?._hasField(tagNumber) ?? false;
   }
 
-  void _clearField(int? tagNumber) {
+  void _clearField(int tagNumber) {
     _ensureWritable();
     final meta = _meta;
-    var fi = _nonExtensionInfo(meta, tagNumber);
+    final fi = _nonExtensionInfo(meta, tagNumber);
+
     if (fi != null) {
-      // clear a non-extension field
-      if (_hasObservers) _eventPlugin!.beforeClearField(fi);
+      assert(tagNumber == fi.tagNumber);
+
+      // Clear a non-extension field
+      final eventPlugin = _eventPlugin;
+      if (eventPlugin != null && eventPlugin.hasObservers) {
+        eventPlugin.beforeClearField(fi);
+      }
       _values[fi.index!] = null;
 
-      if (meta.oneofs.containsKey(fi.tagNumber)) {
-        _oneofCases!.remove(meta.oneofs[fi.tagNumber]);
+      final oneofIndex = meta.oneofs[tagNumber];
+      if (oneofIndex != null) {
+        _oneofCases![oneofIndex] = 0;
       }
 
-      var oneofIndex = meta.oneofs[fi.tagNumber];
-      if (oneofIndex != null) _oneofCases![oneofIndex] = 0;
       return;
     }
 
-    if (_hasExtensions) {
-      var fi = _extensions!._getInfoOrNull(tagNumber);
+    final extensions = _extensions;
+    if (extensions != null) {
+      var fi = extensions._getInfoOrNull(tagNumber);
       if (fi != null) {
-        _extensions!._clearField(fi);
+        extensions._clearField(fi);
         return;
       }
     }
@@ -271,10 +263,11 @@ class _FieldSet {
     final meta = _meta;
     var fi = _nonExtensionInfo(meta, tagNumber);
     if (fi == null) {
-      if (!_hasExtensions) {
+      final extensions = _extensions;
+      if (extensions == null) {
         throw ArgumentError('tag $tagNumber not defined in $_messageName');
       }
-      _extensions!._setField(tagNumber, value);
+      extensions._setField(tagNumber, value);
       return;
     }
 
@@ -337,10 +330,13 @@ class _FieldSet {
 
   /// Sets a non-extended field and fires events.
   void _setNonExtensionFieldUnchecked(BuilderInfo meta, FieldInfo fi, value) {
-    var tag = fi.tagNumber;
-    var oneofIndex = meta.oneofs[tag];
+    final tag = fi.tagNumber;
+    final oneofIndex = meta.oneofs[tag];
     if (oneofIndex != null) {
-      _clearField(_oneofCases![oneofIndex]);
+      final currentOneofTag = _oneofCases![oneofIndex];
+      if (currentOneofTag != null) {
+        _clearField(currentOneofTag);
+      }
       _oneofCases![oneofIndex] = tag;
     }
 
@@ -348,8 +344,9 @@ class _FieldSet {
     // beginning of this method but happens just before the value is set.
     // Otherwise the observers will be notified about 'clearField' and
     // 'setField' events in an incorrect order.
-    if (_hasObservers) {
-      _eventPlugin!.beforeSetField(fi, value);
+    final eventPlugin = _eventPlugin;
+    if (eventPlugin != null && eventPlugin.hasObservers) {
+      eventPlugin.beforeSetField(fi, value);
     }
     _values[fi.index!] = value;
   }
@@ -410,8 +407,7 @@ class _FieldSet {
     assert(fi.isMapField);
 
     if (_isReadOnly) {
-      return PbMap<K, V>.unmodifiable(
-          PbMap<K, V>(fi.keyFieldType, fi.valueFieldType));
+      return PbMap<K, V>(fi.keyFieldType, fi.valueFieldType);
     }
 
     var map = fi._createMapField(_message!);
@@ -488,15 +484,19 @@ class _FieldSet {
     if (value == null) {
       _$check(index, value); // throw exception for null value
     }
-    if (_hasObservers) {
-      _eventPlugin!.beforeSetField(_nonExtensionInfoByIndex(index), value);
+    final eventPlugin = _eventPlugin;
+    if (eventPlugin != null && eventPlugin.hasObservers) {
+      eventPlugin.beforeSetField(_nonExtensionInfoByIndex(index), value);
     }
     final meta = _meta;
     var tag = meta.byIndex[index].tagNumber;
     var oneofIndex = meta.oneofs[tag];
 
     if (oneofIndex != null) {
-      _clearField(_oneofCases![oneofIndex]);
+      final currentOneofTag = _oneofCases![oneofIndex];
+      if (currentOneofTag != null) {
+        _clearField(currentOneofTag);
+      }
       _oneofCases![oneofIndex] = tag;
     }
     _values[index] = value;
@@ -515,21 +515,24 @@ class _FieldSet {
       _unknownFields!.clear();
     }
 
-    if (_hasObservers) {
+    final extensions = _extensions;
+
+    final eventPlugin = _eventPlugin;
+    if (eventPlugin != null && eventPlugin.hasObservers) {
       for (var fi in _infos) {
         if (_values[fi.index!] != null) {
-          _eventPlugin!.beforeClearField(fi);
+          eventPlugin.beforeClearField(fi);
         }
       }
-      if (_hasExtensions) {
-        for (var key in _extensions!._tagNumbers) {
-          var fi = _extensions!._getInfoOrNull(key)!;
-          _eventPlugin!.beforeClearField(fi);
+      if (extensions != null) {
+        for (var key in extensions._tagNumbers) {
+          var fi = extensions._getInfoOrNull(key)!;
+          eventPlugin.beforeClearField(fi);
         }
       }
     }
     if (_values.isNotEmpty) _values.fillRange(0, _values.length, null);
-    if (_hasExtensions) _extensions!._clearValues();
+    extensions?._clearValues();
   }
 
   bool _equals(_FieldSet o) {
@@ -538,14 +541,16 @@ class _FieldSet {
       if (!_equalFieldValues(_values[i], o._values[i])) return false;
     }
 
-    if (!_hasExtensions || !_extensions!._hasValues) {
+    final extensions = _extensions;
+    if (extensions == null || !extensions._hasValues) {
       // Check if other extensions are logically empty.
       // (Don't create them unnecessarily.)
-      if (o._hasExtensions && o._extensions!._hasValues) {
+      final oExtensions = o._extensions;
+      if (oExtensions != null && oExtensions._hasValues) {
         return false;
       }
     } else {
-      if (!_extensions!._equalValues(o._extensions)) return false;
+      if (!extensions._equalValues(o._extensions)) return false;
     }
 
     if (_unknownFields == null || _unknownFields!.isEmpty) {
@@ -613,8 +618,8 @@ class _FieldSet {
     }
 
     // Hash with extension fields.
-    if (_hasExtensions) {
-      final extensions = _extensions!;
+    final extensions = _extensions;
+    if (extensions != null) {
       final sortedByTagNumbers = _sorted(extensions._tagNumbers);
       for (final tagNumber in sortedByTagNumbers) {
         final fi = extensions._getInfoOrNull(tagNumber)!;
@@ -695,15 +700,18 @@ class _FieldSet {
           fi.name == '' ? fi.tagNumber.toString() : fi.name);
     }
 
-    if (_hasExtensions) {
-      _extensions!._info.keys.toList()
+    final extensions = _extensions;
+    if (extensions != null) {
+      extensions._info.keys.toList()
         ..sort()
         ..forEach((int tagNumber) => writeFieldValue(
             _extensions!._values[tagNumber],
             '[${_extensions!._info[tagNumber]!.name}]'));
     }
-    if (_hasUnknownFields) {
-      out.write(_unknownFields.toString());
+
+    final unknownFields = _unknownFields;
+    if (unknownFields != null) {
+      out.write(unknownFields.toString());
     } else {
       out.write(UnknownFieldSet().toString());
     }
@@ -724,43 +732,42 @@ class _FieldSet {
       var value = other._values[fi.index!];
       if (value != null) _mergeField(fi, value, isExtension: false);
     }
-    if (other._hasExtensions) {
-      var others = other._extensions!;
-      for (var tagNumber in others._tagNumbers) {
-        var extension = others._getInfoOrNull(tagNumber)!;
-        var value = others._getFieldOrNull(extension);
+
+    final otherExtensions = other._extensions;
+    if (otherExtensions != null) {
+      for (var tagNumber in otherExtensions._tagNumbers) {
+        var extension = otherExtensions._getInfoOrNull(tagNumber)!;
+        var value = otherExtensions._getFieldOrNull(extension);
         _mergeField(extension, value, isExtension: true);
       }
     }
 
-    if (other._hasUnknownFields) {
-      _ensureUnknownFields().mergeFromUnknownFieldSet(other._unknownFields!);
+    final otherUnknownFields = other._unknownFields;
+    if (otherUnknownFields != null) {
+      _ensureUnknownFields().mergeFromUnknownFieldSet(otherUnknownFields);
     }
   }
 
-  void _mergeField(FieldInfo otherFi, fieldValue, {bool? isExtension}) {
+  void _mergeField(FieldInfo otherFi, fieldValue, {required bool isExtension}) {
     final tagNumber = otherFi.tagNumber;
 
     // Determine the FieldInfo to use.
     // Don't allow regular fields to be overwritten by extensions.
     final meta = _meta;
     var fi = _nonExtensionInfo(meta, tagNumber);
-    if (fi == null && isExtension!) {
+    if (fi == null && isExtension) {
       // This will overwrite any existing extension field info.
       fi = otherFi;
     }
-
-    var mustClone = _isGroupOrMessage(otherFi.type);
 
     if (fi!.isMapField) {
       if (fieldValue == null) {
         return;
       }
       final MapFieldInfo<dynamic, dynamic> f = fi as dynamic;
-      mustClone = _isGroupOrMessage(f.valueFieldType);
       final PbMap<dynamic, dynamic> map =
           f._ensureMapField(meta, this) as dynamic;
-      if (mustClone) {
+      if (_isGroupOrMessage(f.valueFieldType)) {
         PbMap fieldValueMap = fieldValue;
         for (final entry in fieldValueMap.entries) {
           map[entry.key] = (entry.value as GeneratedMessage).deepCopy();
@@ -772,7 +779,7 @@ class _FieldSet {
     }
 
     if (fi.isRepeated) {
-      if (mustClone) {
+      if (_isGroupOrMessage(otherFi.type)) {
         // fieldValue must be a PbList of GeneratedMessage.
         PbList<GeneratedMessage> pbList = fieldValue;
         var repeatedFields = fi._ensureRepeatedField(meta, this);
@@ -788,7 +795,7 @@ class _FieldSet {
     }
 
     if (otherFi.isGroupOrMessage) {
-      final currentFi = isExtension!
+      final currentFi = isExtension
           ? _ensureExtensions()._getFieldOrNull(fi as Extension<dynamic>)
           : _values[fi.index!];
 
@@ -801,7 +808,7 @@ class _FieldSet {
       }
     }
 
-    if (isExtension!) {
+    if (isExtension) {
       _ensureExtensions()
           ._setFieldAndInfo(fi as Extension<dynamic>, fieldValue);
     } else {
@@ -812,7 +819,13 @@ class _FieldSet {
 
   // Error-checking
 
-  /// Checks the value for a field that's about to be set.
+  /// Checks that the message is mutable (not frozen) and the value for the
+  /// field is valid for the field type.
+  ///
+  /// Throws [UnsupportedError] if the message if immutable (frozen).
+  ///
+  /// Throws [ArgumentError] if the field value is not valid for the field
+  /// type. For example, when setting a proto `string` field a Dart `int`.
   void _validateField(FieldInfo fi, var newValue) {
     _ensureWritable();
     var message = _getFieldError(fi.type, newValue);
@@ -836,9 +849,10 @@ class _FieldSet {
   }
 
   bool _hasRequiredExtensionValues() {
-    if (!_hasExtensions) return true;
-    for (var fi in _extensions!._infos) {
-      var value = _extensions!._getFieldOrNull(fi);
+    final extensions = _extensions;
+    if (extensions == null) return true;
+    for (var fi in extensions._infos) {
+      var value = extensions._getFieldOrNull(fi);
       if (!fi._hasRequiredValues(value)) return false;
     }
     return true; // No problems found.
@@ -879,12 +893,14 @@ class _FieldSet {
       }
     }
 
-    if (original._hasExtensions) {
-      _ensureExtensions()._shallowCopyValues(original._extensions!);
+    final originalExtensions = original._extensions;
+    if (originalExtensions != null) {
+      _ensureExtensions()._shallowCopyValues(originalExtensions);
     }
 
-    if (original._hasUnknownFields) {
-      _ensureUnknownFields()._fields.addAll(original._unknownFields!._fields);
+    final originalUnknownFields = original._unknownFields;
+    if (originalUnknownFields != null) {
+      _ensureUnknownFields()._fields.addAll(originalUnknownFields._fields);
     }
 
     _oneofCases?.addAll(original._oneofCases!);