+++ protobuf/CHANGELOG.md
@@ -13,6 +13,11 @@
   getter `length` removed. ([#721])
 * Update library documentation to hide internals, add documentation for public
   types. ([#681])
+* Avoid copying when reading map fields of read-only messages. ([#741])
+* Fix `PbMap._isReadonly` field initialization in `PbMap.unmodifiable`.
+  ([#741])
+* Fix decoding map fields when key or value (or both) fields of a map entry is
+  missing. ([#719], [#745])
 
 [#183]: https://github.com/google/protobuf.dart/issues/183
 [#644]: https://github.com/google/protobuf.dart/pull/644
@@ -25,6 +30,9 @@
 [d94d3f0]: https://github.com/google/protobuf.dart/commit/d94d3f0
 [#721]: https://github.com/google/protobuf.dart/pull/721
 [#681]: https://github.com/google/protobuf.dart/pull/681
+[#741]: https://github.com/google/protobuf.dart/pull/741
+[#719]: https://github.com/google/protobuf.dart/issues/719
+[#745]: https://github.com/google/protobuf.dart/pull/745
 
 ## 2.1.0